requirements.txt
  python-telegram-bot
  requests
  nest_asyncio
  schedule
  